First commit
In this version of the chatbot assistant, the model can only find the necessary answer and print it on the screen when asked questions that are very similar to the questions in the given dataset. If the question expression entered by the user is very different from the question examples in the dataset, the model does not work consistently.